Refactor utilities: move functions to a new utilities module and update import paths
