+++ tests/unit/pipelines/test_utilities.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 from unittest.mock import MagicMock, patch
-from pipelines.flows.utilities import get_secret
+from pipelines.utilities import get_secret
 
 @pytest.fixture
 def mock_secretmanager():
